hopefully final

+++ lab5/gr9sub1/TOPCHEEVA_TAISIA/src/stats_module.c
@@ -14,8 +14,6 @@ MODULE_AUTHOR("tayjie");
 MODULE_DESCRIPTION("System stats module for lab5");
 
 static struct proc_dir_entry *proc_stats;
-
-
 static int count_processes(void) {
     struct task_struct *task;
     int count = 0;
@@ -29,7 +27,6 @@ static int count_processes(void) {
     return count;
 }
 
-
 static ssize_t stats_read(struct file *file, char __user *ubuf, size_t count, loff_t *ppos) {
     char buf[512];
     int len;
@@ -38,16 +35,16 @@ static ssize_t stats_read(struct file *file, char __user *ubuf, size_t count, lo
     if (*ppos > 0)
         return 0;
 
-
     si_meminfo(&si);
+    
 
     len = snprintf(buf, sizeof(buf),
         "Processes: %d\n"
         "Memory Used: %lu MB\n"
-        "System Uptime: %lu seconds\n",
-        count_processes(), 
+        "System Uptime: %u seconds\n", 
+        count_processes(),
         (si.totalram - si.freeram) * si.mem_unit / (1024 * 1024),
-        jiffies_to_msecs(get_jiffies_64()) / 1000
+        (unsigned int)(jiffies_to_msecs(get_jiffies_64()) / 1000)  
     );
 
     if (copy_to_user(ubuf, buf, len)) {
@@ -80,3 +77,4 @@ static void __exit stats_exit(void) {
 
 module_init(stats_init);
 module_exit(stats_exit);
+
